Move agent check to opentelemtery-tracing
