+++ tracing/opentelemetry/pom.xml
@@ -33,6 +33,10 @@
             <artifactId>helidon-common-context</artifactId>
         </dependency>
         <dependency>
+        <groupId>io.helidon.config</groupId>
+        <artifactId>helidon-config</artifactId>
+    </dependency>
+        <dependency>
             <groupId>io.helidon.tracing</groupId>
             <artifactId>helidon-tracing</artifactId>
         </dependency>
@@ -58,5 +62,6 @@
             <artifactId>hamcrest-all</artifactId>
             <scope>test</scope>
         </dependency>
+
     </dependencies>
 </project>
